fix: Resolve compilation errors in filesystem and USB device modules
- Add missing MOUNT_POINT definition to filesystem.h
- Fix header guard in filesystem.h (#ifndef FILESYSTEM_H)
- Fix function signature mismatches between .h and .c files
- Fix format string for time_t (use %lld instead of %ld)
- Update TinyUSB include from tinyusb.h to esp_tinyusb.h
- Add esp_tinyusb dependency to CMakeLists.txt

These fixes resolve all compilation errors reported by Michael.

+++ esp32s3-dualusb-fw/main/filesystem.c
@@ -138,7 +138,7 @@ bool fs_write_test_file(void) {
 
     time_t now = time(NULL);
     fprintf(f, "Test write at %s", ctime(&now));
-    fprintf(f, "Timestamp: %ld\n", now);
+    fprintf(f, "Timestamp: %lld\n", (long long)now);
 
     int ret = fclose(f);
     fs_unlock();
